categorization/requirements.txt
psycopg[binary]
python-dotenv
confluent-kafka
psycopg-pool
Unidecode
